backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue
celery==5.3.6
redis==5.0.1

# File Storage
boto3==1.34.34  # For AWS S3
python-magic==0.4.27  # libmagic is needed on system

# ML/AI
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
opencv-python-headless==4.9.0.80
librosa==0.10.1
scikit-learn==1.4.0
numpy==1.26.3
pillow==10.2.0
transformers>=4.35.0
accelerate>=0.24.0

# Blockchain (optional for Pipeline 2)
# web3==6.15.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
reportlab==4.4.10
